[63] refactor: extract SegmentWord
