+++ src/index.html
@@ -75,7 +75,7 @@
         padding: 0;
         font-family: var(--system-font-stack);
         color: var(--color-text-primary);
-        font-size: 18px;
+        font-size: 16px;
         line-height: 1.5rem;
         background-color: var(--color-background-light);
       }
