+++ src/views/tableview/table-view.js
@@ -7,15 +7,17 @@ import { getLanguageFromFilename } from '../utility/views-common';
 
 import sharedDataViewStyles from '../data/data-view-shared.styles';
 
-const TableViewInfoModalContent = () => html`
-  <div>
-    <p>
-      Displays only the parallel numbers. It is possible to sort the parallels
-      according to their position in the main text, grouped by the text in which
-      they appear and by their length.
-    </p>
-  </div>
-`;
+function TableViewInfoModalContent() {
+  return html`
+    <div>
+      <p>
+        Displays only the parallel numbers. It is possible to sort the parallels
+        according to their position in the main text, grouped by the text in
+        which they appear and by their length.
+      </p>
+    </div>
+  `;
+}
 
 @customElement('table-view')
 export class TableView extends LitElement {
