+++ src/views/textview/text-view-left.js
@@ -5,7 +5,7 @@ import { getFileTextAndParallels } from '../../api/actions';
 import sharedDataViewStyles from '../data/data-view-shared.styles';
 import styles from './text-view-table.styles';
 import { LeftSegmentContainer } from './LeftSegment';
-import { C_HIGHLIGHTED_SEGMENT, C_SELECTED_SEGMENT } from './text-view';
+import { C_HIGHLIGHTED_SEGMENT } from './text-view';
 
 /**
  * TODO:
@@ -105,7 +105,10 @@ export class TextViewLeft extends LitElement {
       active_segment: this.leftActiveSegment,
     });
     this.endOfLeftTextFlag = textleft.length !== 200;
+    console.log({ textleft });
+    console.log('removing duplicates');
     this.textLeft = removeDuplicates(textleft, 'segnr');
+    console.log(this.textLeft);
     this.textLeftBySegNr = {};
     this.textLeft.forEach(
       ({ segnr, parallel_ids }) => (this.textLeftBySegNr[segnr] = parallel_ids)
@@ -176,12 +179,12 @@ export class TextViewLeft extends LitElement {
   }
 
   handleSegmentClick(e) {
-    this.shadowRoot
-      .querySelectorAll(`.${C_SELECTED_SEGMENT}`)
-      .forEach(el => el.classList.remove(C_SELECTED_SEGMENT));
-    this.shadowRoot
-      .querySelectorAll(`.${C_HIGHLIGHTED_SEGMENT}`)
-      .forEach(el => el.classList.remove(C_HIGHLIGHTED_SEGMENT));
+    // this.shadowRoot
+    //   .querySelectorAll(`.${C_SELECTED_SEGMENT}`)
+    //   .forEach(el => el.classList.remove(C_SELECTED_SEGMENT));
+    // this.shadowRoot
+    //   .querySelectorAll(`.${C_HIGHLIGHTED_SEGMENT}`)
+    //   .forEach(el => el.classList.remove(C_HIGHLIGHTED_SEGMENT));
     if (e) {
       this.displayParallels(e.target);
     }
@@ -238,6 +241,7 @@ export class TextViewLeft extends LitElement {
   }
 
   render() {
+    console.log(this.textLeft);
     return html`
       ${this.fetchLoading
         ? html`
