+++ src/views/utility/preprocessing.js
@@ -1,7 +1,8 @@
 // this file contains preprocessing routines used both by text-view and table-view for the
 // text strings coming from the database.
 import { html } from 'lit-element';
-import { C_SELECTED_SEGMENT } from '../textview/text-view';
+
+import { SegmentWord } from '../textview/SegmentWord';
 
 export const preprocessTibetan = currentString => {
   currentString = currentString.replace(/\//g, '|') + ' ';
@@ -75,7 +76,7 @@ export const highlightTextByOffset = (
         }
         colourValues.push(colourValue);
       }
-      let tokenizedResult = tokenizeWords({
+      let tokenizedResult = SegmentWord({
         inputData: textArray[i],
         lang: lang,
         colorValues: colourValues,
@@ -99,10 +100,6 @@ export const colorTable = {
   10: '#FF860D',
 };
 
-function getCooccuranceColor(cooc) {
-  return cooc < 10 ? colorTable[cooc] : colorTable[10];
-}
-
 export function segmentArrayToString(segmentArray) {
   let SegmentRef = segmentArray[0];
   if (segmentArray.length > 1) {
@@ -110,96 +107,6 @@ export function segmentArrayToString(segmentArray) {
     SegmentRef = SegmentRef + `–${parallelArray.slice(-1)[0]}`;
   }
   return SegmentRef;
-}
-
-const wrapWordsInSpan = (
-  selectedSegment,
-  currentColor,
-  rightMode,
-  position,
-  clickFunction,
-  cleanedWord
-) => {
-  if (currentColor == -1) {
-    // prettier-ignore
-    return html`<span 
-        class="word ${selectedSegment}"
-        position="${position}"
-        @click="${clickFunction}">${cleanedWord}</span>`
-  }
-  if (!currentColor || currentColor == 0) {
-    return cleanedWord;
-  }
-  let highlightColor = rightMode
-    ? '#2ECC40'
-    : getCooccuranceColor(currentColor);
-  // prettier-ignore
-  return html`<span 
-        class="word highlight-parallel ${selectedSegment}"
-        style="color:${highlightColor}"
-        position="${position}"
-        @click="${clickFunction}">${cleanedWord}</span>`
-};
-
-export function tokenizeWords({
-  inputData,
-  lang,
-  colorValues,
-  clickFunction = 0,
-  highlightMode = 0,
-  rightMode = false,
-}) {
-  let words = [];
-  let selectedSegment = '';
-  if (colorValues.length > 0) {
-    let splitWords = inputData;
-    if (lang.match(/tib|pli/)) {
-      // this is a small hack to avoid line breaks when a * || combination occurs in ACIP
-      inputData = inputData.replace(/\* \//, '*_/');
-      splitWords = inputData.split(' ');
-    }
-    let position = 0;
-    for (let i = 0; i < splitWords.length; ++i) {
-      let currentColor = colorValues.length >= 0 ? colorValues[i] : 0;
-      if (currentColor !== 0 && highlightMode === 1) {
-        selectedSegment = C_SELECTED_SEGMENT;
-      }
-      let cleanedWord = '';
-      if (lang.match(/tib|pli/)) {
-        cleanedWord = preprocessTibetan(splitWords[i]);
-      } else {
-        cleanedWord = preprocessChineseCharacter(splitWords[i]);
-      }
-      words.push(
-        wrapWordsInSpan(
-          selectedSegment,
-          currentColor,
-          rightMode,
-          position,
-          clickFunction,
-          cleanedWord
-        )
-      );
-      if (lang.match(/tib|pli/)) {
-        position += splitWords[i].length + 1;
-      } else {
-        position += splitWords[i].length;
-      }
-    }
-  } else {
-    if (lang.match(/tib|pli/)) {
-      words = preprocessTibetan(inputData);
-    } else {
-      words = inputData.split('').map(character => {
-        return preprocessChineseCharacter(character);
-      });
-    }
-  }
-  return lang !== 'skt'
-    ? words
-    : html`
-        ${words}<br />
-      `;
 }
 
 export function getLinkForSegmentNumbers(language, segmentnr) {
